pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecosystem-data-simulator"
version = "0.1.0"
description = "Ecosystem Data Simulator: Causal core + Event synthesizer + Intelligence graph"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=0.24.0",
    "pyyaml>=5.4.0",
    "faker>=8.0.0",
    "tqdm>=4.60.0",
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0"
]

[project.optional-dependencies]
dev = [
    "black>=21.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "isort>=5.9.0"
]

[tool.setuptools]
packages = ["ecosystem_sim"]
